+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+### 0.1.0
+- Add [Shape](https://github.com/noah-friedman/obj.GL/blob/v0.1.0/src/Shape) class
+- Add option to choose draw mode in [DrawItem.draw()](https://github.com/noah-friedman/obj.GL/blob/v0.1.0/src/DrawItem/index.ts#L132)
+
 ### 0.0.6
 - Add [.npmignore](https://github.com/noah-friedman/obj.GL/blob/v0.0.6/.npmignore)
 - Reconfigure file structure in published package